+++ scripts/Groovy/lib/entities/CompoundFilter.groovy
@@ -1,0 +1,9 @@
+package lib.entities
+
+import groovy.transform.ToString
+
+@ToString(includeNames = true, includeSuperProperties = true, ignoreNulls = true)
+class CompoundFilter extends SimpleFilter {
+    Operator operator = Operator.AND
+    List<CompoundFilter> filters
+}