+++ scripts/Groovy/lib/entities/Condition.groovy
@@ -1,0 +1,6 @@
+package lib.entities
+
+enum Condition {
+    IN,
+    NIN
+}