+++ scripts/Groovy/lib/entities/Query.groovy
@@ -1,0 +1,10 @@
+package lib.entities
+
+import groovy.transform.ToString
+
+@ToString(includeNames = true, ignoreNulls = true)
+class Query {
+    List<String> select
+    List<Reduction> reductions
+    CompoundFilter where
+}