+++ scripts/Groovy/lib/entities/SimpleFilter.groovy
@@ -1,0 +1,7 @@
+package lib.entities
+
+class SimpleFilter {
+    String field
+    String condition
+    List<Object> values
+}